Refactor `PanGestureManager` to improve compile time
1. Flatten nested types
2. Add more type annotations at `Float.normalized` (ex.
`PanGestureManager.normalize(_:)`)
